+++ src/main.tsx
@@ -1,10 +1,17 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import QRky from './components/QRky.tsx'
+import { BrowserRouter, Routes, Route } from 'react-router-dom'
+import Home from './pages/Home'
+import Changelog from './pages/Changelog'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <QRky />
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/changelog" element={<Changelog />} />
+      </Routes>
+    </BrowserRouter>
   </React.StrictMode>,
 )