+++ src/pages/Home.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import QRky from '../components/QRky';
+
+const Home: React.FC = () => {
+    return <QRky />;
+};
+
+export default Home;